Move interop root to a separate file
