+++ zk_ee/src/common_structs/mod.rs
@@ -4,6 +4,7 @@ pub mod events_storage;
 pub mod history_counter;
 pub mod history_list;
 pub mod history_map;
+pub mod interop_root;
 pub mod logs_storage;
 pub mod new_preimages_publication_storage;
 pub mod proof_data;
